+++ guillotina/db/storages/base.py
@@ -4,9 +4,9 @@ class BaseStorage(object):
 
     _cache_strategy = 'dummy'
     _read_only = False
-    _transaction_strategy = 'merge'
+    _transaction_strategy = 'resolve'
 
-    def __init__(self, read_only=False, transaction_strategy='merge',
+    def __init__(self, read_only=False, transaction_strategy='resolve',
                  cache_strategy='dummy'):
         self._read_only = read_only
         self._transaction_strategy = transaction_strategy